Use save-core to get plugins and resources
### What's done:
* save-cli snapshot version accessor
