+++ save-orchestrator/build.gradle.kts
@@ -1,11 +1,9 @@
 import org.cqfn.save.buildutils.configureJacoco
 import org.cqfn.save.buildutils.configureSpringBoot
+import org.cqfn.save.buildutils.getSaveCliVersion
 
 import de.undercouch.gradle.tasks.download.Download
-import org.codehaus.groovy.runtime.ResourceGroovyMethods
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-import java.net.URL
 
 plugins {
     kotlin("jvm")
@@ -22,20 +20,7 @@ tasks.withType<KotlinCompile> {
 }
 
 tasks.getByName("processResources").finalizedBy("downloadSaveCli")
-val saveCliVersion: String = if (Versions.saveCore.endsWith("SNAPSHOT")) {
-    // try to get the required version of cli
-    findProperty("saveCliVersion") as String? ?: run {
-        // as fallback, use latest release to allow the project to build successfully
-        val latestRelease = ResourceGroovyMethods.getText(
-            URL("https://api.github.com/repos/cqfn/save/releases/latest")
-        )
-        (groovy.json.JsonSlurper().parseText(latestRelease) as Map<String, Any>)["tag_name"].let {
-            (it as String).trim('v')
-        }
-    }
-} else {
-    Versions.saveCore
-}
+val saveCliVersion = getSaveCliVersion()
 tasks.register<Download>("downloadSaveCli") {
     dependsOn("processResources")
     // if required, file can be provided manually
